move post_save signal receivers to tags package
