+++ cyphon/alerts/tests/test_signals.py
@@ -26,16 +26,12 @@ except ImportError:
     from mock import Mock, patch
 
 # third party
-from django.contrib.contenttypes.models import ContentType
-from django.db.models.signals import post_save
-from django.test import TestCase, TransactionTestCase
+from django.test import TestCase
 from testfixtures import LogCapture
 
 # local
 from appusers.models import AppUser
-from alerts.models import Alert, Analysis, Comment
-from alerts.signals import tag_alert, tag_analysis, tag_comment
-from tags.models import TagRelation
+from alerts.models import Comment
 from tests.fixture_manager import get_fixtures
 
 
@@ -113,96 +109,3 @@ class SendCommentNotificationTestCase(TestCase):
                          'An error occurred when sending an email '
                          'notification: (535, \'foobar\')'),
                     )
-
-
-class TagAlertTestCase(TransactionTestCase):
-    """
-    Tests the tag_alert receiver.
-    """
-    fixtures = get_fixtures(['alerts', 'datataggers'])
-
-    def setUp(self):
-        super(TagAlertTestCase, self).setUp()
-        post_save.connect(tag_alert, sender=Alert)
-        self.alert = Alert.objects.get(pk=2)
-
-    def tearDown(self):
-        super(TagAlertTestCase, self).tearDown()
-        post_save.disconnect(tag_alert, sender=Alert)
-
-    def test_old_alert(self):
-        """
-        Tests that an old Alert is tagged.
-        """
-        self.alert.save()
-        self.assertEquals(len(self.alert.associated_tags), 2)
-
-    def test_new_alert(self):
-        """
-        Tests that a new Alert is tagged.
-        """
-        alert = self.alert
-        alert.pk = None
-        alert.save()
-        self.assertEquals(len(alert.associated_tags), 2)
-
-
-class TagAnalysisTestCase(TransactionTestCase):
-    """
-    Tests the tag_comment receiver.
-    """
-    fixtures = get_fixtures(['alerts', 'tags'])
-
-    def setUp(self):
-        super(TagAnalysisTestCase, self).setUp()
-        post_save.connect(tag_analysis, sender=Analysis)
-
-    def tearDown(self):
-        super(TagAnalysisTestCase, self).tearDown()
-        post_save.disconnect(tag_analysis, sender=Analysis)
-
-    def test_tag_analysis(self):
-        """
-        Tests that a Comment is tagged when saved.
-        """
-        analysis = Analysis.objects.get(pk=3)
-        analysis_type = ContentType.objects.get_for_model(Analysis)
-        tag_relations = TagRelation.objects.filter(content_type=analysis_type,
-                                                   object_id=analysis.pk)
-        self.assertEquals(tag_relations.count(), 1)
-        analysis.notes = 'I like cats and dogs.'
-        analysis.save()
-        tag_relations = TagRelation.objects.filter(content_type=analysis_type,
-                                                   object_id=analysis.pk)
-        self.assertEquals(tag_relations.count(), 3)
-
-
-class TagCommentTestCase(TransactionTestCase):
-    """
-    Tests the tag_comment receiver.
-    """
-    fixtures = get_fixtures(['comments', 'tags'])
-
-    def setUp(self):
-        super(TagCommentTestCase, self).setUp()
-        post_save.connect(tag_comment, sender=Comment)
-
-    def tearDown(self):
-        super(TagCommentTestCase, self).tearDown()
-        post_save.disconnect(tag_comment, sender=Comment)
-
-    def test_tag_comment(self):
-        """
-        Tests that a Comment is tagged when saved.
-        """
-        comment = Comment.objects.get(pk=3)
-        comment_type = ContentType.objects.get_for_model(Comment)
-        tag_relations = TagRelation.objects.filter(content_type=comment_type,
-                                                   object_id=comment.pk)
-        self.assertEquals(tag_relations.count(), 0)
-
-        comment.content = 'I like cats and dogs.'
-        comment.save()
-        tag_relations = TagRelation.objects.filter(content_type=comment_type,
-                                                   object_id=comment.pk)
-        self.assertEquals(tag_relations.count(), 2)
