+++ cyphon/tags/signals.py
@@ -20,30 +20,34 @@
 
 # standard library
 import logging
-import smtplib
 
 # third party
+from django.conf import settings
 from django.db.models.signals import post_save
-from django.dispatch import receiver
 
 # local
-from utils.emailutils.emailutils import emails_enabled
-from .models import Comment
-from .services import compose_comment_email
+from alerts.models import Alert, Analysis, Comment
+from .models import DataTagger, Tag
 
 _LOGGER = logging.getLogger(__name__)
 
 
-@receiver(post_save, sender=Comment)
-def send_comment_notification(sender, instance, created, **kwargs):
-    """Email relevant users when a new |Comment| is saved."""
+def tag_alert(sender, instance, created, **kwargs):
+    """Tag a new |Alert|."""
+    DataTagger.objects.process(instance)
 
-    # email relevant users if a comment is created
-    if created and emails_enabled():
-        for user in instance.get_other_contributors():
-            email_message = compose_comment_email(instance, user)
-            try:
-                email_message.send()
-            except smtplib.SMTPAuthenticationError as error:
-                _LOGGER.error('An error occurred when sending an '
-                              'email notification: %s', error)
+
+def tag_analysis(sender, instance, created, **kwargs):
+    """Tag an |Analysis|."""
+    Tag.objects.process(value=instance.notes, obj=instance)
+
+
+def tag_comment(sender, instance, created, **kwargs):
+    """Tag a |Comment|."""
+    Tag.objects.process(value=instance.content, obj=instance)
+
+
+if not settings.TEST:
+    post_save.connect(tag_alert, sender=Alert)
+    post_save.connect(tag_analysis, sender=Analysis)
+    post_save.connect(tag_comment, sender=Comment)
